Finalized all TF features.

All monuments implemented and automated.
TF Monuments homebrew option added.

+++ src/main/java/ti4/helpers/ButtonHelperAbilities.java
@@ -19,6 +19,7 @@ import ti4.ResourceHelper;
 import ti4.discord.interactions.buttons.Buttons;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.whispers.tyris.TyrisAbilityHandler;
 import ti4.discord.interactions.buttons.handlers.faction.homebrew.whispers.zephyrion.ZephyrionBountyHandler;
+import ti4.discord.interactions.buttons.handlers.unit.monuments.TwilightsFallMonumentsButtonHandler;
 import ti4.discord.interactions.routing.ButtonHandler;
 import ti4.game.Game;
 import ti4.game.Planet;
@@ -2153,10 +2154,15 @@ public final class ButtonHelperAbilities {
 
     @ButtonHandler("enterCoexistence_")
     public static void enterCoexistence(String buttonID, ButtonInteractionEvent event, Game game, Player player) {
-        event.getMessage().delete().queue(Consumers.nop(), BotLogger::catchRestError);
         String planet = buttonID.split("_")[1];
         UnitHolder unitHolder = game.getUnitHolderFromPlanet(planet);
         Tile tile = game.getTileFromPlanet(planet);
+        if (TwilightsFallMonumentsButtonHandler.preventsCoexistence(game, tile)) {
+            MessageHelper.sendEphemeralMessageToEventChannel(
+                    event, "Units cannot enter coexistence in The Crown Of Thorns system.");
+            return;
+        }
+        event.getMessage().delete().queue(Consumers.nop(), BotLogger::catchRestError);
         List<Player> playersWithUnitsOnPlanet = ButtonHelper.getPlayersWithUnitsOnAPlanet(game, unitHolder);
         Optional<Player> enemyPlayer = playersWithUnitsOnPlanet.stream()
                 .filter(p -> player != p && !player.isPlayerMemberOfAlliance(p))
